requirements.txt
# Core
pyspark==3.5.1
delta-spark==3.2.0

# Kafka
kafka-python==2.0.2

# Dashboarding
streamlit==1.35.0
altair==5.3.0
wordcloud==1.9.3

# Data manipulation
pandas==2.2.2
pyarrow==15.0.2
